backend/requirements.txt
# FastAPI framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
lxml==4.9.3
html5lib==1.1
fake-useragent==1.3.0
aiohttp==3.8.6

# NLP and text processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
tenacity==8.2.3
pydantic[email]==2.4.2
schedule==1.2.1
aiofiles==23.2.1
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and monitoring
loguru==0.7.2